refactor: explore
